container: move to go-embed for re-exec of C code
Instead of doing this horrible thing where we use
__attribute__((constructor)) to do hook the binary before the golang
runtime runs, let's just use a separate binary that we write out to a
memfd and exec things that way.

This also fixes some horribleness with glibc's setxid gunk hanging for
unclear reasons when statically linked.

This drops support for golang 1.15 in CI, since go-embed is a 1.16-only
feature. Hopefully this won't be too much of a problem :)

Because this uses memfd, it only works on kernels >= 3.17.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>