import: drop chmod removal
This is a spiritual revert of the code in 9f46e4efbd23 ("import: handle
readonly parent directories") (but not the test); since we're now running
things inside a userns, the kernel's "is this thing root" checks pass and
we don't need to do this dance any more. Let's drop it.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>