tar base layers: fix computation of dependency order
The problem here is that we explicitly allow and document stacker:// style
imports for the url: portion of tar: base layers, so people can compose
tarballs in one layer and emit them in other layers.

However, the dependency order calculation did not take into account this
dependency. Let's add it to the computation.

Additionally, we fix a comment on DependencyOrder(), since it *does* in
fact re-order things now (and the comment had some poor grammar...)

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>