overlay: fix re-build of build-only overlay layers
The problem here was in our "do changes exist" detection in
generateLayer(). We just tested if the overlay/ working directory was
empty.

However, on previous builds where there were intermediary build_only
layers but a final OCI output, we would go through and generate an OCI
layer for each of the intermediary build layers.

Then the rebuild would go through, and if none of the contents of these
layers changed in a build, they would be ignored since there were no
changes.

Let's add a flag to indicate "hey, we actually built this earlier, you
should grab the output from there" if there are no changes. Note that an
empty dir really does mean no changes, since if someone deletes stuff (or
changes file metadata, or whatever), overlay will generate a whiteout node
or do a copy_up() to have somewhere to put the metadata, etc.

This is strictly better than passing around whether or not the layer was
actually rebuilt, since we can re-use things in the case where the layer is
smart enough not to actually change anything if it doesn't want to. Maybe a
small optimization, but I believe the code is also smaller as well...

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>