overlay: test that the overlayMetadata structure is unchanged
If the overlay metadata structure changes, it probably implies a bug fix or
some other change in behavior, which previous overlay metadata will not
satisfy (even though they'll unparse correctly, the fields will just have
their "zero" value).

See the comment in TestCacheEntryChanged() for a full explanation, but
basically we should re-build everything if this structure has changed,
which means we should have a test for when it changes so that if people
change it, they know to bump the cache version.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>