unpriv-setup: don't try to resolve username from uid
The problem here is that os/user.LookupId() will fail if a uid isn't in
/etc/passwd:

https://cs.opensource.google/go/go/+/refs/tags/go1.17.2:src/os/user/lookup_unix.go;l=240

which is a reasonably common situation with enterprise deployments, network
filesystems, etc.

However, in all these cases I'm aware of, people are doing unpriv-setup via
sudo, and not passing explicit uids and gids, so we can pull the username
out of sudo as well.

For anyone who *was* explicitly passing uids and gids, they probably know
what --username to pass as well. We could do some backwards compat
LookupId() if username == "" for those folks if they do end up existing.
But for now let's just punt.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>