overlay: cache image config as well as manifest in metadata
The problem here was that since we were grabbing the Manifest from the
mutator without committing, it didn't re-generate the descriptor in
manifest.Config so the DiffIDs would be stale.

We could Commit() the manifest, but then it might be dangling and get
cleaned up on GC (e.g. in the case of build layers which aren't emitted,
our new committed manifest+config won't be referenced by anything).

So instead, let's just cache the config, since we want to save it.

This requires an umoci API change, since the mutator only returns the
ImageConfig, and not the Image itself. Hopefully that will be upstream
soon.

Note that we also have to bump the cache version since there's a change in
the overlay metadata.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>