build: use lxc git master
while we're waiting for build failure fixes to be backported, let's use the
latest tagged version.

it looks like the patch that broke things (59d8a539d106) has been around
for a long time, so this is likely a result of gcc's static analysis
getting smarter, and lxc upstream not using alpine with the latest gcc.

the fix is bdb8aeda721f9262695001a7368bb991527b4b8a which is in lxc master
now.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>