test: use same LXC_BRANCH when building n-1 version
in this caching test we build stacker N-1 to test that we can read the
previous version's cache. this is not LXC dependent, so let's use whatever
the current LXC_BRANCH is built with.

this will save us in cases where a particular commit has an LXC_BRANCH with
bugs (or compiler errors).

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>