storage: return an error if trying to switch storage types
Generally, this indicates a bug somewhere up the stack. For example, when
using shared stacker caches in CI systems, switching back and forth can
cause issues.

While it was not necessarily hard to support switching storage types, it
seems safer to just fail. Users can do the manual clean themselves that
this code does fairly simply, so not much UX is lost :)

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>