container: get rid of re-exec of lxc-usernsexec
This re-exec is kind of a pain, because it means we rely on some reasonable
version of the lxc-usernsexec binary being around. Given that the goal of
all the static linking work was to make it so that we could just copy a
binary around, let's just open code the tiny bit of lxc-usernsexec that we
actually rely on. I tried to upstream a new API for this, but the lxc test
suite kept failing in CI and passing locally for me, with strange errors
that seemed like memory corruption :(.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>