ci: drop version hard coding of golangci-lint
While this is nice because it doesn't randomly break builds, without
somebody monitoring golangci releases, we never get new versions of the
linter and potentially better fixes.

Let's just drop this for now and let the man on the internet break our CI.
If it becomes too painful we can re-enable it.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>