types: unify layer parsing + errors
Stacker layers are "yaml in name only", since we have implemented a bunch
of semantics on top of yaml (and allow different types of things for the
same entry). This was probably a bad idea, like most new software is :). It
has all grown organically and sort of exists piecemeal everywhere
throughout this package. Let's unify all of this parsing so it's all in one
place.

In addition, we absolute+normalize all paths during the parsing of layers,
instead of later. This means we do not have to keep around the
referenceDirectory, which I believe was causing problems in layer cache
comparisons. This also means we don't need all of the ParseFoo() helpers,
since we do everything right after unmarshaling the layers.

Finally, we move all *Layer pointers to Layer references, as in the
previous commit, since we do not intend to edit a user representation of a
layer. This requires a fix to one test which relied on this behavior in a
very roundabout way...

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>