build: disallow duplicate layer names in output
Duplicate layer names via prerequisites or build-multiple invocations will
overwrite each other, causing the output to depend on which is built first.

Additionally, they can cause caching problems if they are based on
different things (this is the bug that 170df8060e62 ("cache: don't hash,
use reflect.DeepEqual() for layer comparisons") was intended to fix; the
additional logging from that commit helped identify the real bug).

So, let's enforce that all layer names are unique across all prerequisites
and matched stacker files after they are parsed.

Now that .build() is a private API, we can just do this in the
BuildMultiple() call path.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>