*: delete the btrfs storage backend
We have not been adding new features to this for some time (instead
choosing to require overlay), and porting it to new umoci APIs will be
painful. It is slower than the overlay backend, which is now reasonably
battle tested, having been heavily used for nearly two years. Let's delete
it.

This drops:

* support from the testsuite matrix
* a bunch of btrfs-only functions from the storage API whose overlay
  implementations were no-ops
* some functions that are no longer called in the oci, squashfs packages

We keep around the storage type detection code, since we'd still like to
refuse to build on backends where btrfs used to be present.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>