squashfs: add verity data
The idea here is that atomfs mounts could be dm-verity verified. That way,
even if someone changes the data, we don't have to do a full read+hash in
order to catch it, we can just catch it with dm-verity on the loopback
device.

Perhaps the most controversial part of this patch is where the verity hash
table ends up living: appended to the end of the squashfs archive. This
might seem like a little bit of a hack, but we can compute where it lives
(see VerityDataLocation()), and could even add auto-detection based on
something else besides mime type if we wanted to. Additionally, there is
precedent here, since the dm-verity tools support storing this data on the
same block device.

There is a gigantic bonus, though: we don't have to augment all existing
tooling with support for some new verity data location. Given that it has
taken us three years and we have yet to successfully land squashfs support
of any form in skopeo, this is a big win IMO.

This patch involves a few different simultaneous changes:

* make LayerType a struct, so that we can add metadata about whether verity
  data is present or not (especially in the cache...). this also requires
  bumping the cache version
* add support for actually generating the verity data via libcryptsetup.
  this requires adding a bunch of gorp to our build env, since
  libcryptsetup's deps are not all distributed as static libraries by
  Alpine
* a slight re-work of how lxc-wrapper is built in order to support the
  above

Finally, we can only retrieve the root hash via an API that was added in
cryptsetup 2.3.0. It's tough to detect this without doing a real verity
format, so let's just catch this case when we run into it, and skip the
unit test in that case.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>