ci: add go 1.18 to CI
note that there *is* motivation for keeping golang 1.17 around:
golangci-lint only runs certain linters with golang 1.18 right now,
because generics broke the AST apparently.

So, let's keep 1.17 around and make sure that those linters still run,
since eventually they'll be re-enabled and this way we won't have to go
through and fix everything when they are.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>