squashfs: rearrange the way we pass ldlibs
this way downstream golang projects don't have to know to pass all this
extra nonsense in their builds, they get it for "free".

we do it in this strange way with static_build because newer versions of
cryptsetup (in particular, the latest version, which we're building to
statically link against) depend on json-c, which is not in a lot of repos
(e.g. ubuntu 20.04).

so in order to work around this, we only expose these flags when we
actually need them, when doing a static build.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>