fix(bind): interpret bind defined as type []map[string]string
As per the type definition of Layer, bind is of type Bind which is a
struct with fields Source and Dest.

But when we do UnmarshalYAML for type Bind, we are assuming that we
would always get the Bind as a []string and not
[]map[interface{}]interface{} (when making a call to
getStringOrStringSlice).

This PR fixes the getStringOrStringSlice to consider Bind defined
as a []map[string]string and returns appropriate bind string
in that case.

Fixes https://github.com/project-stacker/stacker/issues/301

Signed-off-by: Darshil Parikh <darshil.parikh@gmail.com>