feat: Switch MaybeRunInUserns to MaybeRunInNamespace, confine root.
2 things here:
 * Add functionality to call lxc-wrapper 'usernsnexec' or 'nsexec'
   based upon whether or not the caller is uid 0.
 * Track whether or not we have entered a namespace
   by referencing/adding STACKER_REAL_UID environment variable.
   If that variable is found, then MayberunInNamespace will
   simply execute the command without changing a namespace.

'stacker internal-go cp' is used internally by 'stacker grab'.
It is run inside a container. Previosly it avoided being put in
a namespace simply because the Geteuid() was 0.  But now we put
root in a namespace also, so that won't work.  The change here
(cmd/main.go) is to make 'stacker internal-go cp' be identified as
something that should not be run in a namespace.

Last, change the name MayberunInUserns to MaybeRunInNamespace
to reflect the fact that it won't sometimes it runs in
a namespace that is not a usernamespace.

Signed-off-by: Scott Moser <scmoser@cisco.com>