feat: Use squashfuse to "extract" squash filesystems.
Use squashfuse to mount a squash filesystem rather than extracting
it with unsquashfs.  This is nice because :
 * don't need unsquashfs binary (but we do need squashfuse)
 * unsquashfs running unprivledged will error and behave differently
   in different versions.  Such an example is when extracting device
   nodes.
 * faster use than extracting.

Given a mountpoint of 'path/to/my/mountpoint', the squashfuse debug
output is logged to 'path/to/my/.mountpoint-squashfs.log'.

Note that we are nto currently cleaning up any squashfuse processes,
but rather relying on the fact that we should be executing in a pid
and mount namespace, so the kernel will clean them all up.

Note we could support a "real root" option here to mount
via kernel mount rather than squashfuse.

Signed-off-by: Scott Moser <scmoser@cisco.com>