test: add code coverage support
With golang 1.20.x, we can build a binary with coverage support.

"
Cover

Go 1.20 supports collecting code coverage profiles for programs
(applications and integration tests), as opposed to just unit tests.

To collect coverage data for a program, build it with go build's -cover
flag, then run the resulting binary with the environment variable
GOCOVERDIR set to an output directory for coverage profiles. See the
'coverage for integration tests' landing page for more on how to get
started. For details on the design and implementation, see the proposal.
"

Signed-off-by: Ramkumar Chinchani <rchincha@cisco.com>