feat: Use kernel mount of squashfs if possible.
Try to mount a squashfs image with a kernel mount.
This should be faster than squashfuse mount, but will only
work if we are host root.

Environment variable STACKER_ALLOW_SQUASHFS_KERNEL_MOUNTS can
control whether or not kernel mounts can be tried.
It defaults to 'true', but if it is set to 'false', then no
attempt will be made.

Signed-off-by: Scott Moser <scmoser@cisco.com>