feat: Remove use of anuvu/squashfs and libsquashfs1
The superblock.go comes from
https://github.com/diskfs/go-diskfs/blob/master/filesystem/squashfs/

diskfs is MIT licensed and this one Apache, so copy is fine.

The project doesn't expose any of the superblock internals (even size),
so we can't just go import it.

The only change from diskfs's current tip is to replace fmt.Errorf with
errors.Errorf.

Signed-off-by: Scott Moser <smoser@brickies.net>