fix: build.yaml run section needs set -e, and add -x
Its been a long time that this has been broken.  The run sections
in build.yaml were executing shell scripts without '-e' and thus
they would not exit on failure.

The end result was that if something went amuck you had to parse
through long error logs for what that was.

Also
 * Add '-x' to sh to make it more obvious where things failed.
 * Add the po4a package, which is used as part of the xz build.

Signed-off-by: Scott Moser <smoser@brickies.net>