feat: Support spaces in bom author/vendor and cmd exec inside lxc. (#487)
* fix: Use arrays for execution rather than scalar string.

This changes the interface for container.Execute to be an
array of strings rather than a string, and then changes
the internal lxc-wrapper pass arguments provided to it
through to the lxc 'start' api call.

It ultimately allows us to safely call programs with spaces,
quotes or other "odd" characters without first serializing
them in go and relying on lxc start api call to unserialize
them properly.  lxc.start would ultimately end up just
split the lxc.execute.cmd on spaces.

Signed-off-by: Scott Moser <smoser@brickies.net>

* test: Change bom test to contain spaces

The bom test case previously had Author of 'bom-test',
which was not realiastic and was intentional to avoid
exposing the bug with spaces.

Signed-off-by: Scott Moser <smoser@brickies.net>

* refactor: Bind stacker binary into container in SetupBuildContainerConfig

Previously there were 3 individual calls to bind stacker into
tools/static-stacker.  This just changes that to 1 place, so
after calling SetupBuildContainerConfig you can use stacker inside.

Signed-off-by: Scott Moser <smoser@brickies.net>

---------

Signed-off-by: Scott Moser <smoser@brickies.net>