fix: Get nightly workflow running, cleanup build workflow.
The build workflow was being 'used' by callers in ci and release.
Here also make 'nightly' a caller.

Additionally,
 * add 'slow-test' input to build.yaml, defaulting to true.
   Set that to false in ci.yaml so we do not run the slow tests
   on all PRs.
   slow-test value gets put into environment variable SLOW_TEST
   which is read by the test harness.
 * make build-id a scalar, not an array.
 * move build-id out of matrix (it is a constant) and reference
   it as an input instead.
 * Provide default values for 'privilege-level' and 'go-version'
   in build.yaml so the callers do not have to specify those.

Note that we still *do* specify the privilege-level and go-version
in a release build.  This is, I believe, because we want a consistent
thing that gets restored in the cache.

It is unfortunate, but as a result, we aren't testing with unpriv
during a release build.

Signed-off-by: Scott Moser <smoser@brickies.net>