fix: use a version of umoci with fix for failing chmod
There is a PR against umoci#main to avoid trying to chmod +r when we
don't need to.  This will avoid trying to chmod +r on a mountpoint, which
is the root cause of #450.

While we can and should also forward port our umoci dep to point at
current main, we should not do that as part of this PR, as that will
risk confusion if there are regressions from just the fwd port.

Closes #450

Signed-off-by: Serge Hallyn <serge@hallyn.com>