fix: Improvements to yaml.Unmarshal for types.Imports and add some go tests
There is one change here, and that is to break if a stacker yaml
has an import that is a dictionary type.  The content in an 'import'
is now only supported as a string or a list, not a dictionary.

For example, these are allowed:

 * import: /path/to/file

 * import:
    - /file
    - /file2
    - path: /file3

But this is not:

 * import:
     path: /file3

Improvements here:
 * Drop duplicate code blocks from getImportFromInterface
   There were 2 identical codeblocks, one with a type convert to
   map[string]interface{} and one with type convert to
   map[interface{}]interface{}.  That was present to handle the dict
   case.
 * Add some golang tests to for unmarshalling.
 * Type check the values of all string Import fields, rather than just
   converting them to a string with 'fmt.Sprintf("%v")'

Signed-off-by: Scott Moser <smoser@brickies.net>