fix: Improve bind imports, support lower case 'source' in yaml, add go test
Entries stacker.yaml's "binds" are of type Bind.
When imported from yaml, support was present for these three things:
1. ascii art specifying source and dest ("source -> dest")
2. source only ("source")
3. map with 'Source' and 'Dest'.

Support for '3' is dropped, and replaced by map with 'source' and
'dest'.  'Source' and 'Dest' are inconsistent with the rest of stacker.yaml.
Note that importing from json still does support capital letter variants
as the golang json library is case insensitive.

Improvements to the bind yaml and json support generally rely more
heavily on the library to do the lifting.

Signed-off-by: Scott Moser <smoser@brickies.net>