feat: use squashfuse ready notifier if available (#528)
When using squasfuse, check whether it supports -o notifypipe. If it
does, use it instead our manual checking of the mountpoint inode. The
notification mechanism is a better alternative to the existing polling
approach. If it's not available, then use the old mechanism.

This feature is supported starting from squasfuse version 0.5.0, see [1]
for details.

[1] https://github.com/vasi/squashfuse/issues/49

Co-Developed-by: Serge Hallyn <serge@hallyn.com>

Signed-off-by: Ariel Miculas <amiculas@cisco.com>