feat: Support legacy 'import' directive and use 'imports' for future
This changes the behavior of the 'import' target to behave as it did
before the breaking change that moved imports into /stacker/imports/

So now, if the stacker file uses 'import', then imports will
be placed in /stacker.  If the stacker file uses 'imports' (plural)
then they will be placed in /stacker/imports.

What we actually get in both cases is a consistent set of binds being
done into a different "base".  Either /.stacker (legacy) or /stacker (new).

   stacker -> <base>/bin/stacker
   imports -> <base>/imports
   runscript -> <base>/imports/.stacker-run.sh
   artifacts -> <base>/artifacts

In the legacy case, the imports are also bind-mounted into /stacker

Signed-off-by: Scott Moser <smoser@brickies.net>