test: add busybox image, use it to mostly replace centos image
This is basically the add of BUSYBOX image and then using
that instead of centos image.

Moving to busybox from centos required a few changed assumptions:

 * drop -P from grep - busybox's grep did not support -P. instead
   use posix groups ([:space:])
 * don't use long --format for 'stat' - busybox stat didn't like
   this, so just use '-c'.
 * mount - /usr/bin/mount -> /bin/mount - busybox image only has
   /bin, not /usr/bin/mount
 * /bin/bash - busybox image does not have bin/bash

The tests "squashfs empty change no layer" and
"tar empty change no layer" do not work with the busybox image.
There are layers created, and they have this content:

    $ tar tvf oci/blobs/sha256/acff571f8cf*b03f45
    tar: Removing leading `/' from member names
    drwxr-xr-x 0/0               0 2023-11-13 14:06 /
    drwxr-xr-x 0/0               0 2023-11-13 14:06 etc/
    ---------- 0/0               0 2023-11-13 14:06 etc/resolv.conf
    drwxr-xr-x 0/0               0 2023-11-13 14:06 proc/
    drwxr-xr-x 0/0               0 2023-11-13 14:06 sys/

The content there are results of stacker.  I don't have a good way
to avoid them at the moment.

I've also left the bom tests in place with the centos image.
Those would have had to have been adjusted and probably
would not work with busybox at the moment.

Signed-off-by: Scott Moser <smoser@brickies.net>