test: Fix test failure due to .coverage changing during tar creation.
This seems to be a transient failure, but is legitimate issue.

When coverage binary is being used, this test case could fail with:

    #  + mkdir /output
    # + tar -cf /output/installer-iso.tar -C /stacker .
    # tar: ./.coverage: file changed as we read it

The /stacker/.coverage file was being updated by the stacker binary
itself.  This is just a race condition based on when the file gets
written to and when it was being collected by tar.

The change is just to only have tar collect the .cpio files.

Signed-off-by: Scott Moser <smoser@brickies.net>