feat: Support legacy 'import' directive and use 'imports' for future (#543)
* feat: Support legacy 'import' directive and use 'imports' for future

This changes the behavior of the 'import' target to behave as it did
before the breaking change that moved imports into /stacker/imports/

So now, if the stacker file uses 'import', then imports will
be placed in /stacker.  If the stacker file uses 'imports' (plural)
then they will be placed in /stacker/imports.

What we actually get in both cases is a consistent set of binds being
done into a different "base".  Either /.stacker (legacy) or /stacker (new).

   stacker -> <base>/bin/stacker
   imports -> <base>/imports
   runscript -> <base>/imports/.stacker-run.sh
   artifacts -> <base>/artifacts

In the legacy case, the imports are also bind-mounted into /stacker

Signed-off-by: Scott Moser <smoser@brickies.net>

* test: replace import with imports, add one explicit legacy import test

This just changes all the existing tests to test 'imports' rather
than 'import'.  Then it adds one test for 'import' explicitly
and within the same stacker file.

Test test/cache.bats test "can read previous version's cache" is skipped
as the old version can't build a stacker file in the test because
it uses 'imports'.

Signed-off-by: Scott Moser <smoser@brickies.net>

---------

Signed-off-by: Scott Moser <smoser@brickies.net>